Added exceptions for storage component

+++ src/Bloomkit/Core/Storage/Storage.php
@@ -5,6 +5,8 @@ namespace Bloomkit\Core\Storage;
 use InvalidArgumentException;
 use Bloomkit\Core\Storage\Adapter\StorageAdapterInterface;
 use Bloomkit\Core\Storage\Utils\Utils;
+use Bloomkit\Core\Storage\Exceptions\FileNotFoundException;
+use Bloomkit\Core\Storage\Exceptions\FileExistsException;
 
 class Storage implements StorageInterface
 {
@@ -46,8 +48,10 @@ class Storage implements StorageInterface
     {
         $path = Utils::normalizePath($path);
         $newpath = Utils::normalizePath($newpath);
-        $this->assertPresent($path);
-        $this->assertAbsent($newpath);
+        if(!$this->has($path))
+        	throw new FileNotFoundException();
+       	if($this->has($newpath))
+       		throw new FileExistsException();
 
         return $this->getAdapter()->copy($path, $newpath);
     }
@@ -69,7 +73,8 @@ class Storage implements StorageInterface
     public function delete($path)
     {
         $path = Utils::normalizePath($path);
-        $this->assertPresent($path);
+        if(!$this->has($path))
+        	throw new FileNotFoundException();
 
         return $this->getAdapter()->delete($path);
     }
@@ -129,7 +134,8 @@ class Storage implements StorageInterface
     public function getMetadata($path)
     {
         $path = Utils::normalizePath($path);
-        $this->assertPresent($path);
+        if(!$this->has($path))
+        	throw new FileNotFoundException();
 
         return $this->getAdapter()->getMetadata($path);
     }
@@ -140,7 +146,8 @@ class Storage implements StorageInterface
     public function getMimetype($path)
     {
         $path = Utils::normalizePath($path);
-        $this->assertPresent($path);
+        if(!$this->has($path))
+        	throw new FileNotFoundException();
         if ((!$object = $this->getAdapter()->getMimetype($path)) || !array_key_exists('mimetype', $object)) {
             return false;
         }
@@ -154,7 +161,8 @@ class Storage implements StorageInterface
     public function getSize($path)
     {
         $path = Utils::normalizePath($path);
-        $this->assertPresent($path);
+        if(!$this->has($path))
+        	throw new FileNotFoundException();
         if ((!$object = $this->getAdapter()->getSize($path)) || !array_key_exists('size', $object)) {
             return false;
         }
@@ -168,7 +176,8 @@ class Storage implements StorageInterface
     public function getTimestamp($path)
     {
         $path = Utils::normalizePath($path);
-        $this->assertPresent($path);
+        if(!$this->has($path))
+        	throw new FileNotFoundException();
         if ((!$object = $this->getAdapter()->getTimestamp($path)) || !array_key_exists('timestamp', $object)) {
             return false;
         }
@@ -182,7 +191,8 @@ class Storage implements StorageInterface
     public function getVisibility($path)
     {
         $path = Utils::normalizePath($path);
-        $this->assertPresent($path);
+        if(!$this->has($path))
+        	throw new FileNotFoundException();
         if ((!$object = $this->getAdapter()->getVisibility($path)) || !array_key_exists('visibility', $object)) {
             return false;
         }
@@ -264,7 +274,9 @@ class Storage implements StorageInterface
     public function read($path)
     {
         $path = Utils::normalizePath($path);
-        $this->assertPresent($path);
+        if(!$this->has($path))
+        	throw new FileNotFoundException();
+        
         if (!($object = $this->getAdapter()->read($path))) {
             return false;
         }
@@ -278,7 +290,9 @@ class Storage implements StorageInterface
     public function readStream($path)
     {
         $path = Utils::normalizePath($path);
-        $this->assertPresent($path);
+        if(!$this->has($path))
+        	throw new FileNotFoundException();
+        
         if (!$object = $this->getAdapter()->readStream($path)) {
             return false;
         }
@@ -293,8 +307,10 @@ class Storage implements StorageInterface
     {
         $path = Utils::normalizePath($path);
         $newpath = Utils::normalizePath($newpath);
-        $this->assertPresent($path);
-        $this->assertAbsent($newpath);
+        if(!$this->has($path))
+        	throw new FileNotFoundException();
+        if($this->has($newpath))
+        	throw new FileExistsException();
 
         return (bool) $this->getAdapter()->rename($path, $newpath);
     }
@@ -305,7 +321,8 @@ class Storage implements StorageInterface
     public function setVisibility($path, $visibility)
     {
         $path = Utils::normalizePath($path);
-        $this->assertPresent($path);
+        if(!$this->has($path))
+        	throw new FileNotFoundException();
 
         return (bool) $this->getAdapter()->setVisibility($path, $visibility);
     }
@@ -317,7 +334,8 @@ class Storage implements StorageInterface
     {
         $path = Utils::normalizePath($path);
         $config = $this->prepareConfig($config);
-        $this->assertPresent($path);
+        if(!$this->has($path))
+        	throw new FileNotFoundException();
 
         return (bool) $this->getAdapter()->update($path, $contents, $config);
     }
@@ -332,7 +350,9 @@ class Storage implements StorageInterface
         }
         $path = Utils::normalizePath($path);
         $config = $this->prepareConfig($config);
-        $this->assertPresent($path);
+        if(!$this->has($path))
+        	throw new FileNotFoundException();
+        
         Utils::rewindStream($resource);
 
         return (bool) $this->getAdapter()->updateStream($path, $resource, $config);
@@ -344,7 +364,8 @@ class Storage implements StorageInterface
     public function write($path, $contents, array $config = [])
     {
         $path = Utils::normalizePath($path);
-        $this->assertAbsent($path);
+        if($this->has($path))
+        	throw new FileExistsException();
         $config = $this->prepareConfig($config);
 
         return (bool) $this->getAdapter()->write($path, $contents, $config);
@@ -359,7 +380,8 @@ class Storage implements StorageInterface
             throw new InvalidArgumentException(__METHOD__.' expects argument #2 to be a valid resource.');
         }
         $path = Utils::normalizePath($path);
-        $this->assertAbsent($path);
+        if($this->has($path))
+        	throw new FileExistsException();
         $config = $this->prepareConfig($config);
         Utils::rewindStream($resource);
